feat: Simplify CLI architecture and add deep mode to batch processing
- Remove complex batch flows, keep only batch_simple.py
- Streamline CLI to three modes: single, deep, batch
- Add deep mode support to batch processing with --deep parameter
- Fix BatchGeneratorAgent API compatibility issues
- Update documentation for simplified interface
- Enhance Prefect flow orchestration with proper error handling
