+++ batch_results/05_Какова_основная_тема_статьи_Ксавье_Ле_Торривеллека.md
@@ -1,0 +1,53 @@
+# Результат исследования
+
+**Запрос:** Какова основная тема статьи Ксавье Ле Торривеллека (2007) о нациестроительстве башкир и татар?
+
+**Агент:** Unknown
+
+**Модель:** gpt-5
+
+**Глубина исследования:** 2
+
+## Ответ
+
+
+
+## Источники
+
+18. [Татары и государственная национальная политика в ...](https://historicalethnology.org/wp-content/uploads/2025/09/%D0%98%D0%AD_3_2025_453-471.pdf)
+36. [2000 X2009](https://www.abimperio.net/indexai.pdf)
+38. [Историческая этнология Historical Ethnology](https://historicalethnology.org/wp-content/uploads/2025/09/%D0%98%D0%AD_3_2025-2.pdf)
+41. [(PDF) Состояние и особенности «российского ислама»](https://www.academia.edu/5049767/%D0%A1%D0%BE%D1%81%D1%82%D0%BE%D1%8F%D0%BD%D0%B8%D0%B5_%D0%B8_%D0%BE%D1%81%D0%BE%D0%B1%D0%B5%D0%BD%D0%BD%D0%BE%D1%81%D1%82%D0%B8_%D1%80%D0%BE%D1%81%D1%81%D0%B8%D0%B9%D1%81%D0%BA%D0%BE%D0%B3%D0%BE_%D0%B8%D1%81%D0%BB%D0%B0%D0%BC%D0%B0_)
+23. [Библиография трудов И.К.Загидуллина.](https://tatarica.org/download_file/view/11835)
+44. [Москва Горячая линия – Телеком 2018](https://static.iea.ras.ru/news/Svoi_i_chuzhie.pdf)
+37. [Свои и чужие. Метаморфозы идентичности на востоке ...](https://www.academia.edu/38712093/%D0%A1%D0%B2%D0%BE%D0%B8_%D0%B8_%D1%87%D1%83%D0%B6%D0%B8%D0%B5_%D0%9C%D0%B5%D1%82%D0%B0%D0%BC%D0%BE%D1%80%D1%84%D0%BE%D0%B7%D1%8B_%D0%B8%D0%B4%D0%B5%D0%BD%D1%82%D0%B8%D1%87%D0%BD%D0%BE%D1%81%D1%82%D0%B8_%D0%BD%D0%B0_%D0%B2%D0%BE%D1%81%D1%82%D0%BE%D0%BA%D0%B5_%D0%B8_%D0%B7%D0%B0%D0%BF%D0%B0%D0%B4%D0%B5_%D0%95%D0%B2%D1%80%D0%BE%D0%BF%D1%8B)
+37. [2000 X2009 - Ab Imperio](https://www.yumpu.com/en/document/view/36800334/2000-x2009-ab-imperio)
+45. [Mezhkonfessional'nye otnosheniya II sbornik.indd](https://dokumen.pub/download/b14b8f1ce04a83d891b51a0783d9421d.html)
+10. [Этнопанорама. 2014. № 1–2 (33). Научно- ...](https://www.osu.ru/sites/niisu/docs/etnopanorama_2014_1-2.pdf)
+11. [(PDF) The state and stakes of Islam “from” Russia](https://www.academia.edu/5189963/The_state_and_stakes_of_Islam_from_Russia)
+12. [Carl Beck Papers](https://carlbeckpapers.pitt.edu/ojs/index.php/cbp/article/view/133/134)
+13. [Central Eurasian Studies Review](https://centraleurasia.org/wp-content/uploads/2018/04/Vol.6_no.12_fulljournal_CESR.pdf)
+14. [(PDF) The Islamic Revival in Russia](https://www.academia.edu/4315435/The_Islamic_Revival_in_Russia)
+15. [Central Eurasian Studies Review. - Brian Glyn Williams](https://www.brianglynwilliams.com/pdfs/CESR_06_12.pdf)
+16. [Threads of Empire: Loyalty and Tsarist Authority in ...](https://dokumen.pub/threads-of-empire-loyalty-and-tsarist-authority-in-bashkiria-15521917-0253019338-9780253019332.html)
+17. [Central Eurasian Reader 9783112400388, 9783879973477](https://dokumen.pub/central-eurasian-reader-9783112400388-9783879973477.html)
+43. [научно-публицистический журнал](https://www.osu.ru/sites/niisu/docs/etnopanorama_2012_3-4.pdf)
+40. [МАРТ–АПРЕЛЬ 2020](https://new.ras.ru/upload/iblock/075/z2owqok80x2kgz76b1livj6niwlt0rdl.pdf)
+20. [Xavier University](https://www.xavier.edu/)
+21. [Current Students - Xavier University](https://www.xavier.edu/students/)
+22. [Tuition and Fees - Xavier University](https://www.xavier.edu/tuition-fees/index)
+23. [Academics | Xavier University](https://www.xavier.edu/academics/index)
+24. [Visit Campus - Undergraduate Admission | Xavier University](https://www.xavier.edu/undergraduate-admission/visit/index)
+25. [Athletics | Xavier University](https://www.xavier.edu/athletics/index)
+26. [Employee Hub | Xavier University](https://www.xavier.edu/employees/index)
+27. [Xavier Admission - Admission | Xavier University](https://www.xavier.edu/admission/index)
+28. [Application Management - Xavier University](https://admit.xavier.edu/apply/)
+29. [Giving to Xavier | Xavier University](https://www.xavier.edu/giving/index)
+30. [Татары и Башкиры: История в Зеркальном Отражении ...](https://muse.jhu.edu/article/561085)
+31. [Ab Imperio-2/2007 - Project MUSE](https://muse.jhu.edu/issue/30951)
+36. [Скачать Ab imperio 2007 №02 [PDF]](https://j.twirpx.link/file/1385937/)
+34. [На правах рукописи - Диссертационные советы УдГУ](https://dissovet.udsu.ru/doc/dissertations/000629-synopsis-%D0%B0%D0%B2%D1%82%D0%BE%D1%80%D0%B5%D1%84%D0%B5%D1%80%D0%B0%D1%82%20%D0%93%D0%B0%D1%80%D0%B8%D1%84%D1%83%D0%BB%D0%BB%D0%B8%D0%BD.pdf)
+38. [Muslims of Central Eurasia in the 21st Century_book-2024](https://mobile.ikiacademy.org/uploads/0-1-202409121726134932.pdf)
+35. [Ab Imperio – 2007 - Project MUSE](https://muse.jhu.edu/pub/178/article/561249/summary)
+39. [(PDF) Russian Ethnography of the 1850s and the Ethos ...](https://www.academia.edu/7321252/Russian_Ethnography_of_the_1850s_and_the_Ethos_of_Civilizing_Mission_The_Case_of_the_Literary_Expedition_RUS_)
+42. [Историческая этнология Historical Ethnology](https://historicalethnology.org/wp-content/uploads/2025/09/%D0%98%D0%AD_3_2025.pdf)
