+++ batch_results/20250918_003156_Первое описание башкир у Ибн Фадлана датировка.md
@@ -1,0 +1,36 @@
+# Первое описание башкир у Ибн Фадлана: датировка
+
+*Created: 2025-09-18 00:31:56*
+
+Краткий ответ: 922 год. В ходе посольства Аббасидского халифата к волжским булгарам Ахмад ибн Фадлан прошёл через земли башкир и оставил их описание в своей «Записке» (Рисале) именно в 922 году [25], [12], [16].
+
+Обоснование и источники:
+- Энциклопедические справочники о самом Ибн Фадлане указывают, что его поездка состоялась в 921–922 гг., а описания народов, включая башкир, относятся к отчёту по этой миссии; таким образом, фиксация башкир датируется 922 годом прибытия посольства на Волгу [25].
+- Публикация Исторического факультета МГУ с полным текстом «Путешествия на Волгу» уточняет хронологию: посольство вышло из Багдада в 921 г., прибыло в Булгарию в мае 922 г.; в тексте приводятся хиджрские датировки, соответствующие 922–923 гг. н.э. [12], [28].
+- Издание на портале «Восточная литература» (Vostlit) даёт полный перевод «Книги Ахмеда Ибн‑Фадлана о путешествии на Волгу», где рассматриваемые эпизоды описания кочевых племён, в том числе башкир (башкорт), относятся к пути миссии весной 922 года [16].
+- Исследовательская заметка Института степи (Оренбург) реконструирует по тексту и астрономическим данным маршрут и график: встреча с башкирами датируется 3 мая 922 г., что конкретизирует внутри 922 года и подтверждает общий вывод [20].
+- Дополнительный обзор в проекте «История.РФ» прямо формулирует: «В описании Ахмеда ибн‑Фадлана (922 г.)» — приводя цитаты об обычаях башкир и относя это к 922 году [23].
+
+Итог: Первый зафиксированный у Ибн Фадлана рассказ об этнониме и обычаях башкир относится к 922 году [25], [12], [16], [20], [23].
+
+- [24] Ученые о предшественниках современных башкир - https://bashmusic.net/ru/blog/bash-orttar/uchenye-o-predshestvennikakh-sovremennykh-bashkir
+- [21] Башкиры - https://ru.wikipedia.org/wiki/%D0%91%D0%B0%D1%88%D0%BA%D0%B8%D1%80%D1%8B
+- [3] Этнография башкир — Википедия - https://ru.wikipedia.org/wiki/%D0%AD%D1%82%D0%BD%D0%BE%D0%B3%D1%80%D0%B0%D1%84%D0%B8%D1%8F_%D0%B1%D0%B0%D1%88%D0%BA%D0%B8%D1%80
+- [27] Первое письменное упоминание о башкирах принадлежит путешественнику ... - https://vk.com/wall-319544_378241
+- [5] Этногенез башкир - kurultai.ru - https://kurultai.ru/ru/content/999-etnogenez-bashkir/
+- [22] Первое письменное упоминание о башкирах | История Уфы и Башкортостана ... - https://smartik.ru/ufa/post/142602083
+- [7] Башкиры (этноним) - https://ru.wikipedia.org/wiki/%D0%91%D0%B0%D1%88%D0%BA%D0%B8%D1%80%D1%8B_(%D1%8D%D1%82%D0%BD%D0%BE%D0%BD%D0%B8%D0%BC)
+- [23] Башкиры: жизнь в движении. Часть 1: прийти к России - https://histrf.ru/read/articles/bashkiry-zhizn-v-dvizhienii-chast-1-priiti-k-rossii
+- [26] Филолог и историк античности С.Я. Лурье ... - https://www.facebook.com/ngsjournal/posts/%D1%84%D0%B8%D0%BB%D0%BE%D0%BB%D0%BE%D0%B3-%D0%B8-%D0%B8%D1%81%D1%82%D0%BE%D1%80%D0%B8%D0%BA-%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%BD%D0%BE%D1%81%D1%82%D0%B8-%D1%81%D1%8F-%D0%BB%D1%83%D1%80%D1%8C%D0%B5-%D1%81%D1%87%D0%B8%D1%82%D0%B0%D0%BB-%D1%87%D1%82%D0%BE-%D0%BF%D1%80%D0%B5%D0%B4%D1%88%D0%B5%D1%81%D1%82%D0%B2%D0%B5%D0%BD%D0%BD%D0%B8%D0%BA%D0%B8-%D1%81%D0%BE%D0%B2%D1%80%D0%B5%D0%BC%D0%B5%D0%BD%D0%BD%D1%8B%D1%85-%D0%B1%D0%B0%D1%88/1110715840498358/
+- [10] Восточные авторы о башкирском обществе - https://urgaza.ru/library-portal/articles/291/1947/
+- [25] Ибн Фадлан — Википедия - https://ru.wikipedia.org/wiki/%D0%98%D0%B1%D0%BD_%D0%A4%D0%B0%D0%B4%D0%BB%D0%B0%D0%BD
+- [28] Ибн-Фадлан. Путешествие на Волгу - https://www.hist.msu.ru/ER/Etext/fadlan.htm
+- [13] Ибн Фадлан: Рисала - История России в датах - https://rushrono.ru/istochniki_vos_018.html
+- [14] «ЗАПИСКА» («РИСАЛА») ИБН ФАДЛАНА И ОСОБЕННОСТИ РАННЕГО ЭТАПА ... - https://cyberleninka.ru/article/n/zapiska-risala-ibn-fadlana-i-osobennosti-rannego-etapa-formirovaniya-volzhskih-bulgar-x-vek
+- [15] путь Ахмеда ибн-Фадлана. - Kitaptar - https://kitaptar.bashkort.org/files/put_ahmeda_ibn_fadlana._ufa._2010.pdf
+- [16] Книга Ахмеда Ибн-Фадлана о его путешествии на Волгу ... - https://www.vostlit.info/Texts/rus16/Fadlan/text.phtml
+- [17] Точная копия «Записки» («Рисала») арабского писателя и путешественника ... - https://vk.com/wall-220809031_13640
+- [29] Башкиры в 922 году. Записки экспедиции Ахмеда ибн ... - https://vk.com/wall103336706_6562
+- [19] «ЗАПИСКА» («РИСАЛА») ИБН ФАДЛАНА И ОСОБЕННОСТИ РАННЕГО ЭТАПА ... - https://www.chuvsu.ru/nauka-i-innovaczii/vestnik-chuvashskogo-universiteta/metadannye-k-statyam/vestnik-chuvashskogo-universiteta-20212-%E2%84%96-2-tema-vypuska-istoricheskie-nauki/zapiska-risala-ibn-fadlana-i-osobennosti-rannego-etapa-formirovaniya-volzhskih-bulgar-x-vek/
+- [20] Маршрут И График Движения Ибн Фадлана По Степям От Устюрта До Волги (X ... - https://orensteppe.org/content/marshrut-i-grafik-dvizheniya-ibn-fadlana-po-stepyam-ot-ustyurta-do-volgi-x-vek
+- [30] История Башкортостана - https://zinref.ru/000_uchebniki/02800_logika/011_lekcii_raznie_39/2962.htm
