+++ sgr_deep_research/flows/__init__.py
@@ -1,6 +1,6 @@
 """Prefect flows for SGR Deep Research agents."""
 
 from .research_flow import research_flow
-from .batch_ultra_simple import batch_create_flow, batch_run_flow
+from .batch_simple import batch_simple_flow
 
-__all__ = ["research_flow", "batch_create_flow", "batch_run_flow"]
+__all__ = ["research_flow", "batch_simple_flow"]
